+++ app/interactors/approve_static_invite.rb
@@ -1,0 +1,6 @@
+# Approving static invite
+class ApproveStaticInvite
+  include Interactor::Organizer
+
+  organize CreateStaticMember, UpdateStaticInvite
+end